- Updated slade.pk3 with more game configurations. Things for Doom, Heretic, Hexen and Strife are complete. - Added "height" field for things in game configurations. It'll be useful whenever 3D mode is implemented, and could also be used for error checking (e.g.: ceiling too low for things). - Added "translation" to thing configurations, and a number of standard translations. Especially useful for Strife. - ThingInfoOverlay and MapCanvas with things_drawtype == 2 will use translations now. The preview sprites therefore look like they do in-game.
git-svn-id: http://mancubus.net/svn/hosted/slade/branches/map_editor@1052 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/SLADEMap.h
@@ -77,11 +77,11 @@ public:
 	MapLine*	getLine(unsigned index);
 	MapSector*	getSector(unsigned index);
 	MapThing*	getThing(unsigned index);
-	int			nVertices() { return vertices.size(); }
-	int			nLines() { return lines.size(); }
-	int			nSides() { return sides.size(); }
-	int			nSectors() { return sectors.size(); }
-	int			nThings() { return things.size(); }
+	size_t		nVertices() { return vertices.size(); }
+	size_t		nLines() { return lines.size(); }
+	size_t		nSides() { return sides.size(); }
+	size_t		nSectors() { return sectors.size(); }
+	size_t		nThings() { return things.size(); }
 
 	// Map structure indices
 	int		vertexIndex(MapVertex* v);
